+++ Tests/Dot-source-ib/Dot-test.ps1
@@ -74,7 +74,7 @@ Push-Location function:
 assert (!(Test-Path Write-Warning))
 
 # expected public functions
-$OK = 'Add-BuildTask,Assert-Build,Assert-BuildEquals,Get-BuildError,Get-BuildFile,Get-BuildProperty,Get-BuildSynopsis,Get-BuildVersion,Invoke-BuildExec,New-BuildJob,Test-BuildAsset,Use-BuildAlias,Write-Build'
+$OK = 'Add-BuildTask,Assert-Build,Assert-BuildEquals,Get-BuildError,Get-BuildFile,Get-BuildProperty,Get-BuildSynopsis,Invoke-BuildExec,New-BuildJob,Test-BuildAsset,Use-BuildAlias,Write-Build'
 $KO = (Get-ChildItem *-Build* -Name | Sort-Object) -join ','
 assert ($OK -ceq $KO) "Unexpected functions:
 OK: [$OK]